Only append node once, rather than once per block
Previously there was no tracking in the `Gemfile#append_before_first_block!`
method to make sure that the node was appended only once, rather than once
before each block in the Gemfile. This fixes that, and also makes sure that
every dumped output has a trailing newline, the latter because each *block*
also has a trailing newline, and we want to be consistent about whether or not
we do that for the whole file too.

Fixes #1